+++ src/common/Smi.Common/Options/GlobalOptions.cs
@@ -88,6 +88,8 @@ namespace Smi.Common.Options
         public ProcessDirectoryOptions ProcessDirectoryOptions { get; set; }
         public DeadLetterReprocessorOptions DeadLetterReprocessorOptions { get; set; }
 
+        public IsIdentifiableOptions IsIdentifiableOptions { get; set; }
+
         #endregion
 
         public static string GenerateToString(object o)
@@ -107,6 +109,22 @@ namespace Smi.Common.Options
         {
             return GenerateToString(this);
         }
+    }
+
+    public class IsIdentifiableOptions : ConsumerOptions
+    {
+        /// <summary>
+        /// The full name of the classifier you want to run
+        /// </summary>
+        public string ClassifierType { get; set; }
+
+        /// <summary>
+        /// The root location in which subfolders must exist containing all data files required by any classifiers
+        /// (typically 1 sub-directory per classifier)
+        /// </summary>
+        public string DataDirectory { get; set; }
+
+        public ProducerOptions IsIdentifiableProducerOptions {get; set;}
     }
 
     [UsedImplicitly]