+++ src/microservices/Microservices.IsIdentifiable/Rules/RuleSet.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microservices.IsIdentifiable.Rules
+{
+    public class RuleSet
+    {
+        public IsIdentifiableRule[] BasicRules { get; set; }
+        public SocketRule[] SocketRules { get; set; }
+    }
+}