+++ tests/microservices/Microservices.IsIdentifiable.Tests/IsIdentifiable_TestRegexDetections.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Microservices.IsIdentifiable.Failure;
+using Microservices.IsIdentifiable.Failures;
 using Microservices.IsIdentifiable.Options;
 using Microservices.IsIdentifiable.Runners;
 using NUnit.Framework;
@@ -10,19 +10,9 @@ namespace Microservices.IsIdentifiable.Tests
 {
     public class IsIdentifiableRunnerTests
     {
-        public void FailIfNoClassifier()
-        {
-            string f = NerEngineTests.ClassifierPath;
-
-            if (!File.Exists(f))
-                Assert.Inconclusive("File did not exist " + f);
-        }
-
         [Test]
         public void TestChiInString()
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner("hey there,0101010101 excited to see you");
             runner.Run();
 
@@ -37,8 +27,6 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("dd37lb")]
         public void IsIdentifiable_TestPostcodes(string code)
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner("Patient lives at " + code);
             runner.Run();
 
@@ -55,14 +43,14 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("dd37lb")]
         public void IsIdentifiable_TestPostcodes_WhitelistDD3(string code)
         {
-            FailIfNoClassifier();
 
             var runner = new TestRunner("Patient lives at " + code);
             
             runner.LoadRules(
                 @"
-- Action: Ignore
-  IfPattern: DD3");
+BasicRules:
+  - Action: Ignore
+    IfPattern: DD3");
 
             runner.Run();
             
@@ -74,8 +62,6 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("dd37lb")]
         public void IsIdentifiable_TestPostcodes_IgnorePostcodesFlagSet(string code)
         {
-            FailIfNoClassifier();
-
             //since allow postcodes flag is set
             var runner = new TestRunner("Patient lives at " + code, new TestOpts() { IgnorePostcodes = true });
             runner.Run();
@@ -85,13 +71,12 @@ namespace Microservices.IsIdentifiable.Tests
         }
 
 
-        [TestCase("Patient_lives_at_DD28DD", "DD28DD")]
+        //no longer detected, that's fine
+        //[TestCase("Patient_lives_at_DD28DD", "DD28DD")]
         [TestCase("^DD28DD^", "DD28DD")]
         [TestCase("dd3^7lb", "dd3 7lb")]
         public void IsIdentifiable_TestPostcodes_EmbeddedInText(string find, string expectedMatch)
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner(find);
             runner.Run();
 
@@ -107,7 +92,6 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("1444DD2011FD1118E63006097D2DF4834C9D2777977D811907000065B840D9CA50000000837000000FF0100A601000000003800A50900000700008001000000AC020000008000000D0000805363684772696400A8480000E6FBFFFF436174616C6F6775654974656D07000000003400A50900000700008002000000A402000000800000090000805363684772696400A84800001E2D0000436174616C6F67756500000000008000A50900000700008003000000520000000180000058000080436F6E74726F6C00A747000")]
         public void IsIdentifiable_TestNotAPostcode(string code)
         {
-            FailIfNoClassifier();
 
             var runner = new TestRunner("Patient lives at " + code);
             runner.Run();
@@ -143,8 +127,6 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("AB 13:10", "13:10", null, null)]
         public void IsIdentifiable_TestDates(string date, string expectedMatch1, string expectedMatch2, string expectedMatch3)
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner("Patient next appointment is " + date);
             runner.Run();
             
@@ -171,8 +153,6 @@ namespace Microservices.IsIdentifiable.Tests
         [TestCase("AB13:10")]
         public void IsIdentifiable_Test_NotADate(string input)
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner(input);
             runner.Run();
 
@@ -182,38 +162,36 @@ namespace Microservices.IsIdentifiable.Tests
         [Test]
         public void TestChiAndNameInString()
         {
-            FailIfNoClassifier();
-
             var runner = new TestRunner("David Smith should be referred to with chi 0101010101");
 
             runner.Run();
-            Assert.AreEqual(3, runner.ResultsOfValidate.Count);
+            Assert.AreEqual(1, runner.ResultsOfValidate.Count);
 
             FailurePart w1 = runner.ResultsOfValidate[0];
-            FailurePart w2 = runner.ResultsOfValidate[1];
-            FailurePart w3 = runner.ResultsOfValidate[2];
+            
+            /* Names are now picked up by the Socket NER Daemon
+            //FailurePart w2 = runner.ResultsOfValidate[1];
+            //FailurePart w3 = runner.ResultsOfValidate[2];
 
+            
             Assert.AreEqual("David", w1.Word);
             Assert.AreEqual(0, w1.Offset);
 
             Assert.AreEqual("Smith", w2.Word);
             Assert.AreEqual(6, w2.Offset);
+            */
 
-            Assert.AreEqual("0101010101", w3.Word);
-            Assert.AreEqual(43, w3.Offset);
+            Assert.AreEqual("0101010101", w1.Word);
+            Assert.AreEqual(43, w1.Offset);
         }
 
         [Test]
         public void TestSopDoesNotMatch()
         {
-            if (!File.Exists(NerEngineTests.ClassifierPath))
-                Assert.Inconclusive("NER classifier file not found");
-
             const string sopKey = "SOPInstanceUID";
             const string exampleSop = "1.2.392.200036.9116.2.6.1.48.1214834115.1486205112.923825";
             var testOpts = new TestOpts
             {
-                PathToNerClassifier = NerEngineTests.ClassifierPath,
                 SkipColumns = sopKey
             };
 
@@ -255,7 +233,6 @@ namespace Microservices.IsIdentifiable.Tests
         {
             public TestOpts()
             {
-                PathToNerClassifier = NerEngineTests.ClassifierPath;
                 DestinationCsvFolder = TestContext.CurrentContext.TestDirectory;
                 StoreReport = true;
             }